Update SoapOperationJpaController.java: getSoapOperationCount() returns a long type now.

+++ src/docprocessing/rpc/soap/persistence/SoapOperationJpaController.java
@@ -202,14 +202,14 @@ public class SoapOperationJpaController implements Serializable {
         }
     }
 
-    public int getSoapOperationCount() {
+    public long getSoapOperationCount() {
         EntityManager em = getEntityManager();
         try {
             CriteriaQuery cq = em.getCriteriaBuilder().createQuery();
             Root<SoapOperation> rt = cq.from(SoapOperation.class);
             cq.select(em.getCriteriaBuilder().count(rt));
             Query q = em.createQuery(cq);
-            return ((Long) q.getSingleResult()).intValue();
+            return (Long) q.getSingleResult();
         } finally {
             em.close();
         }
